Clean up and renaming
